requirements.txt
pytest
pytest-mock
pyyaml
dnspython
